+++ updater/software/CCleaner.cs
@@ -48,13 +48,13 @@ namespace updater.software
         /// <summary>
         /// publisher name for signed executables
         /// </summary>
-        private const string publisherX509 = "CN=Piriform Software Ltd, O=Piriform Software Ltd, L=London, C=GB, SERIALNUMBER=08235567, OID.1.3.6.1.4.1.311.60.2.1.3=GB, OID.2.5.4.15=Private Organization";
+        private const string publisherX509 = "CN=Gen Digital Inc., O=Gen Digital Inc., L=Tempe, S=Arizona, C=US";
 
 
         /// <summary>
         /// expiration date of certificate
         /// </summary>
-        private static readonly DateTime certificateExpiration = new(2024, 12, 4, 23, 59, 59, DateTimeKind.Utc);
+        private static readonly DateTime certificateExpiration = new(2027, 5, 21, 23, 59, 59, DateTimeKind.Utc);
 
 
         /// <summary>
@@ -66,13 +66,13 @@ namespace updater.software
         {
             var signature = new Signature(publisherX509, certificateExpiration);
             var installer = new InstallInfoExe(
-                "https://download.ccleaner.com/ccsetup625.exe",
+                "https://download.ccleaner.com/ccsetup626.exe",
                 HashAlgorithm.SHA256,
-                "e963abf025714d9cba73ccc42a8b8759dc622db08b11e61eb91983a46f73ecc7",
+                "90d47ae79a4c1a7c4b7effb1ec45036a514bc816cf9cf8dfe313b8c74ff4b119",
                 signature,
                 "/S");
             return new AvailableSoftware("CCleaner",
-                "6.25",
+                "6.26",
                 "^CCleaner+$",
                 "^CCleaner+$",
                 // CCleaner uses the same installer for 32 and 64-bit.